requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
Django==3.0.7
django-crispy-forms==1.9.1
django-form-tags==2.0
django-formtools==2.2
django-otp==0.9.3
django-phonenumber-field==3.0.1
django-svg-templatetag==1.0.1
django-two-factor-auth==1.12.1
django-webpush==0.3.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
http-ece==1.1.0
idna==2.10
phonenumbers==8.12.6
Pillow==7.1.2
py-vapid==1.7.1
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
pywebpush==1.9.4
qrcode==6.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
twilio==6.44.0
urllib3==1.25.9
